refactor: extract console renderers
