+++ cli/core/console/__init__.py
@@ -1,0 +1,4 @@
+from cli.core.console.base import console
+from cli.core.console.renderers.banner import show_banner
+
+__all__ = ["console", "show_banner"]
