+++ cli/core/console/base.py
@@ -1,0 +1,3 @@
+from rich.console import Console
+
+console = Console(highlight=False)
